Feat:TI1 - Task 1-13 Done

+++ sipil/src/main/java/apap/tugas/sipil/controller/PenerbanganController.java
@@ -4,6 +4,7 @@ import apap.tugas.sipil.model.PilotModel;
 import apap.tugas.sipil.model.PilotPenerbanganModel;
 import apap.tugas.sipil.repository.PenerbanganDb;
 import apap.tugas.sipil.service.PenerbanganService;
+import apap.tugas.sipil.service.PilotPenerbanganService;
 import apap.tugas.sipil.service.PilotService;
 import apap.tugas.sipil.model.PenerbanganModel;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +13,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -19,6 +22,12 @@ import java.util.NoSuchElementException;
 public class PenerbanganController {
     @Autowired
     private PenerbanganService penerbanganService;
+
+    @Autowired
+    private PilotPenerbanganService pilotPenerbanganService;
+
+    @Autowired
+    private PilotService pilotService;
 
     @GetMapping("/")
     private String home(){
@@ -38,13 +47,34 @@ public class PenerbanganController {
             Model model){
         PenerbanganModel penerbangan = penerbanganService.getPenerbanganById(id);
         List<PilotPenerbanganModel> listBertugas = penerbangan.getListPilotPenerbangan();
+        List<PilotModel> listSemuaPilot = pilotService.getListPilot();
+        PilotPenerbanganModel tempGabungan = new PilotPenerbanganModel();
+        tempGabungan.setPenerbangan(penerbangan);
 
+        model.addAttribute("listSemuaPilot", listSemuaPilot);
         model.addAttribute("penerbangan", penerbangan);
+        model.addAttribute("tempGabungan", tempGabungan);
         model.addAttribute("listBertugas", listBertugas);
         model.addAttribute("adaTugas", !listBertugas.isEmpty());
 
 
         return "view-penerbangan";
+    }
+
+    @PostMapping("/penerbangan/{id}/pilot/tambah")
+    private String tambahPilotPenerbangan(
+            @PathVariable(value= "id") Long id,
+            @ModelAttribute PenerbanganModel penerbangan,
+            @ModelAttribute PilotPenerbanganModel tempGabungan,
+            Model model){
+
+        tempGabungan.setTanggal_penugasan(LocalDate.now());
+        pilotPenerbanganService.addPilotPenerbangan(tempGabungan);
+
+        model.addAttribute("nama", tempGabungan.getPilot().getNama());
+        model.addAttribute("kode", tempGabungan.getPenerbangan().getKode());
+
+        return "penerbangan-add-pilot";
     }
 
     @GetMapping("/penerbangan/tambah")
@@ -71,9 +101,18 @@ public class PenerbanganController {
             if (id == null) {
                 return "error-null-input";
             }
-            PenerbanganModel penerbangan = penerbanganService.getPenerbanganById(id);
-            penerbanganService.deletePenerbangan(penerbangan);
-            return "delete-penerbangan";
+            PenerbanganModel temp = penerbanganService.getPenerbanganById(id);
+            if (temp.getListPilotPenerbangan().isEmpty()){
+                PenerbanganModel penerbangan = penerbanganService.getPenerbanganById(id);
+                String kode = penerbangan.getKode();
+                penerbanganService.deletePenerbangan(penerbangan);
+                model.addAttribute("kode",kode);
+                return "delete-penerbangan";
+            }
+            else{
+                return "delete-penerbangan-gagal";
+            }
+
         }
         catch (NumberFormatException e){
             return "error-input";
